pkg/components/util: use Helm post-renderer instead of Manifests field
As Manifests filed in Chart object is a custom field, added to our fork
of Helm. Now Helm offers post-renderer feature, which allows to further
customize the manfiests after Helm's templating engine and before the
installation of the manifests.

This allows us to use post-renderer feature and drop one of maintained
custom patches, which means less maintenance for us.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>