ks8util: Utility function to update Namespace metadata.
This commit updates labels and annotations of the existing namespace by
first querying the namespace for existing labels and annotations.

The function removes the labels and annotations `lokomotive.kinvolk.io`
added by Lokomotive and retains the user labels and annotations.

Merges the user labels and annotations with the newly provided ones.

Signed-off-by: Imran Pochi <imran@kinvolk.io>